Address Ebuka's feedback
